+++ hive/server/condenser_api/cursor.py
@@ -11,6 +11,17 @@ def last_month():
     """Get the date 1 month ago."""
     return datetime.now() + relativedelta(months=-1)
 
+async def get_post_id(db, author, permlink):
+    """Given an author/permlink, retrieve the id from db."""
+    sql = ("SELECT id FROM hive_posts WHERE author = :a "
+           "AND permlink = :p AND is_deleted = '0' LIMIT 1")
+    return await db.query_one(sql, a=author, p=permlink)
+
+async def get_child_ids(db, post_id):
+    """Given a parent post id, retrieve all child ids."""
+    sql = "SELECT id FROM hive_posts WHERE parent_id = :id AND is_deleted = '0'"
+    return await db.query_col(sql, id=post_id)
+
 async def _get_post_id(db, author, permlink):
     """Get post_id from hive db."""
     sql = "SELECT id FROM hive_posts WHERE author = :a AND permlink = :p"
