+++ hive/server/condenser_api/methods.py
@@ -4,7 +4,7 @@ from functools import wraps
 
 import hive.server.condenser_api.cursor as cursor
 from hive.server.condenser_api.objects import load_posts, load_posts_reblogs
-from hive.server.condenser_api.common import (
+from hive.server.common.helpers import (
     ApiError,
     return_error_info,
     valid_account,
@@ -12,9 +12,7 @@ from hive.server.condenser_api.common import (
     valid_tag,
     valid_offset,
     valid_limit,
-    valid_follow_type,
-    get_post_id,
-    get_child_ids)
+    valid_follow_type)
 
 # pylint: disable=too-many-arguments,line-too-long,too-many-lines
 
@@ -103,7 +101,7 @@ async def get_content(context, author: str, permlink: str):
     db = context['db']
     valid_account(author)
     valid_permlink(permlink)
-    post_id = await get_post_id(db, author, permlink)
+    post_id = await cursor.get_post_id(db, author, permlink)
     if not post_id:
         return {'id': 0, 'author': '', 'permlink': ''}
     posts = await load_posts(db, [post_id])
@@ -117,9 +115,9 @@ async def get_content_replies(context, author: str, permlink: str):
     db = context['db']
     valid_account(author)
     valid_permlink(permlink)
-    parent_id = await get_post_id(db, author, permlink)
+    parent_id = await cursor.get_post_id(db, author, permlink)
     if parent_id:
-        child_ids = await get_child_ids(db, parent_id)
+        child_ids = await cursor.get_child_ids(db, parent_id)
         if child_ids:
             return await load_posts(db, child_ids)
     return []
